Refactor endpoint reply context helpers
